refactor(ColorPicker): slice CustomColorPickers into sub components, hooks, utils
