+++ src/components/molecules/ColorPicker/types.ts
@@ -23,3 +23,16 @@ export interface RGBA extends RGB {
  * The preferred color syntax format to display in the input fields.
  */
 export type ColorFormat = "rgb" | "hex";
+
+/**
+ * Internal representation of a color in the HSV (Hue, Saturation, Value) color space.
+ * This format is mathematically required to accurately calculate 2D palette coordinates.
+ */
+export interface IHsvColor {
+    /** The hue degree on the color wheel (Range: 0 - 360). */
+    hue: number;
+    /** The saturation level (Range: 0 - 1). */
+    saturation: number;
+    /** The brightness/value level (Range: 0 - 1). */
+    value: number;
+}
